devex: Update document client in web-api to no longer require a protocol argument. Fix imports in web-api
