+++ shared/jest-shared.config.ts
@@ -16,10 +16,6 @@ const config: Config = {
     '!src/tools/**/*.ts',
     '!src/test/**/*.ts',
     '!src/**/*_.ts',
-    '!src/persistence/sqs/deleteMessage.ts',
-    '!src/persistence/sqs/getMessages.ts',
-    '!src/persistence/messages/*.ts',
-    '!src/persistence/dynamo/**/*.ts',
     '!src/business/utilities/documentGenerators/**/*.ts',
   ],
   coverageDirectory: './coverage',
