+++ shared/src/business/useCases/serveCaseToIrs/serveCaseToIrsInteractor.ts
@@ -256,7 +256,6 @@ const generateNoticeOfReceipt = async ({
         .getDocument({
           applicationContext,
           key: clinicLetterKey,
-          protocol: 'S3',
           useTempBucket: false,
         });
     }
