+++ web-api/jest-unit.config.ts
@@ -20,13 +20,17 @@ const config: Config = {
     '!src/app-local.{js,ts}',
     '!src/app-public-local.{js,ts}',
     '!src/getDocumentGenerators.ts',
+    '!src/persistence/sqs/deleteMessage.ts',
+    '!src/persistence/sqs/getMessages.ts',
+    '!src/persistence/messages/*.ts',
+    '!src/persistence/dynamo/**/*.ts',
   ],
   coverageDirectory: './coverage',
   coverageProvider: 'babel',
   coverageThreshold: {
     global: {
-      branches: 97,
-      functions: 99,
+      branches: 96,
+      functions: 96,
       lines: 99,
       statements: 99,
     },
