+++ web-client/src/persistence/s3/getDocument.test.ts
@@ -27,74 +27,13 @@ describe('getDocument', () => {
 
     const result = await getDocument({
       applicationContext,
-    } as any);
+      docketNumber,
+      key,
+    });
 
     expect((getPdfFromUrl as jest.Mock).mock.calls[0][0]).toMatchObject({
       url: mockPdfUrl,
     });
     expect(result).toEqual({ name: 'mockfile.pdf' });
-  });
-
-  it('calls S3.getObject when S3 protocol is set', async () => {
-    applicationContext.getStorageClient.mockReturnValue({
-      getObject: jest.fn().mockReturnValue({
-        promise: () => ({ Body: null }),
-      }),
-    });
-    await getDocument({
-      applicationContext,
-      docketNumber,
-      key,
-      protocol: 'S3',
-      useTempBucket: false,
-    });
-
-    expect(applicationContext.getStorageClient().getObject).toHaveBeenCalled();
-  });
-
-  it('retrieves from the temp bucket when S3 protocol is set and useTempBucket is true', async () => {
-    const tempBucketName = 'tempBucket';
-    applicationContext.getTempDocumentsBucketName.mockReturnValue(
-      tempBucketName,
-    );
-    applicationContext.getStorageClient().getObject.mockReturnValue({
-      promise: () =>
-        Promise.resolve({
-          Body: '',
-        }),
-    });
-
-    await getDocument({
-      applicationContext,
-      docketNumber,
-      key,
-      protocol: 'S3',
-      useTempBucket: true,
-    });
-
-    expect(
-      applicationContext.getStorageClient().getObject,
-    ).toHaveBeenCalledWith({ Bucket: tempBucketName, Key: key });
-  });
-
-  it('retrieves from the documents bucket by default when S3 protocol is set', async () => {
-    applicationContext.getStorageClient().getObject.mockReturnValue({
-      promise: () =>
-        Promise.resolve({
-          Body: '',
-        }),
-    });
-
-    await getDocument({
-      applicationContext,
-      docketNumber,
-      key,
-      protocol: 'S3',
-      useTempBucket: false,
-    });
-
-    expect(
-      applicationContext.getStorageClient().getObject,
-    ).toHaveBeenCalledWith({ Bucket: 'DocumentBucketName', Key: key });
   });
 });
